Get rid of experiment with fluid containers ... v2.0 still fixed width
For now.  The top masthead photo is tricky ... what to do once we grow beyond 1200px!  So, forget about it for now and just make do with 1170px for the time being.  Maybe in the future I can have one more fixed width breakpoint and use another photo (like the Emanuel sign on the bricks or something) to expand the masthead a bit more.  We'll see.