Fix a bug in Masthead/Top Nav Partial
The container `<div>` surrounding the navigation bar content had a class that removed some padding and expanded the width for wider total viewport widths (class "layout-v2"). I think this was a holdover from prior experimentation with responsive behavior at wider viewport widths (beyond 1200px).  Fortunately the version deployed to "production" dd not have this defect, but it still existed in the source code.