Tweaking Newsletter Page Layout and Styling
This is a WIP, but I kind of made it live early.  It's mostly good; now we get the thumbnail above the fold on mobile and tablet viewport widths.  It has an issue at narrower mobile (below 500px) where month/year links will wrap for no apparent reason (because the width is artificially constrained).  Hopefully when I replace the widths with accordions, it'll be better.  Progress, not perfection.